New Changes
- optimize the "Report" class to separate complete from another modules
- put some classes and functions in "modules"
- change some names of files
- Delete some epty lines
